Round Manager 1.0
Added going to "North, East, South or West" at the end of the round. (Need to add text to show end of round and where to go)
Added gems that fall from the sky at the end of the round. Maximum of 15 gems. Currently all the same model and value, but will add varying values once more models are in
Deletes all living gems once changing rounds